+++ R/logit.R
@@ -1,6 +1,7 @@
 ##' Logit function
 ##' @param x a number to transform
 ##' @seealso invlogit
+##' @return A numeric
 ##' @export
 
 logit <- function(x) log(x/(1-x))
@@ -8,5 +9,6 @@ logit <- function(x) log(x/(1-x))
 ##' Inverse logit function
 ##' @param x a number to transform
 ##' @seealso logit
+##' @return A numeric
 ##' @export
 invlogit <- function(x) 1/(1+exp(-x))