+++ NEWS.md
@@ -1,3 +1,12 @@
-# 0.0.2
+# NMcalc 0.0.4
+## New features
+* `CVlnorm()` is a new function CV of log-normal dist baed on omega parameters.
+
+# NMcalc 0.0.3
+## New features
+* `quantbin()` supports specification of cutoff probabilities using the
+  `probs` argument.
+
+# NMcalc 0.0.2
 ## New features
 * `logit()` and `invlogit()` are new functions.